requirements.txt
# Core dependencies
pandas>=1.5.0
pyodbc>=4.0.35
psycopg2-binary>=2.9.5
sqlalchemy>=1.4.0
pyyaml>=6.0

# Monitoring and logging dependencies
flask>=2.3.0
psutil>=5.9.0

# Optional: For email alerts (uncomment if needed)
# smtplib (built-in)
# email (built-in) 